\docType{package}
\name{pollster}
\alias{pollster}
\alias{pollster-package}
\title{Huffpost Pollster API}
\description{
This pacakge provides an R interface to the Huffington Post
Pollster API. Pollster provides programmatic access to
opinion polls collected by the Huffington Post.
}
\details{
See \url{http://elections.huffingtonpost.com/pollster/api}
for more details on the API.
}

