% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_tbl.R
\name{pollster_chart_data}
\alias{pollster_chart_data}
\title{Pollster Charts (table)}
\usage{
pollster_chart_data(slug)
}
\arguments{
\item{slug}{The slug of the chart, e.g. "012-general-election-romney-vs-obama.csv".}
}
\value{
A data frame with the chart data.
}
\description{
This function returns the poll data from a Pollster chart in
a convenient tabular form.
}
\examples{
\dontrun{
chart <- pollster_chart_data()
chart
}
}

