% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/house-effects.R
\name{pollster_house_effects}
\alias{pollster_house_effects}
\title{House Effects}
\usage{
pollster_house_effects(slug)
}
\arguments{
\item{slug}{The slug of the chart}
}
\value{
A data frame
}
\description{
For each chart, Pollster estimates house effects for each pollster, which is roughly the average difference
between the polls of a pollster and the average of all the polls. This
is not officially part of the Pollster API.
}
\examples{
\dontrun{
  pollster_house_effects("2016-general-election-trump-vs-clinton")
}
}

