\name{pollstR-package}
\alias{pollstR-package}
\alias{pollstR}
\docType{package}
\title{R client for Pollster API}
\description{Extract polling information and charts from Huffington Post's Pollster API.}
\author{
Thomas J. Leeper, based on some initial work by Drew Linzer.

Maintainer: Thomas J. Leeper <thosjleeper@gmail.com>
}
\references{\url{http://elections.huffingtonpost.com/pollster/api}}
\keyword{ package }
%\seealso{}
%\examples{}
