% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts-csv.R
\name{pollstr_chart_tbl}
\alias{pollstr_chart_tbl}
\title{Pollster Charts (table)}
\usage{
pollstr_chart_tbl(slug)
}
\arguments{
\item{slug}{The slug of the chart, e.g. "012-general-election-romney-vs-obama.csv".}
}
\value{
A data frame with the chart data.
}
\description{
This function returns the poll data from a Pollster chart in
a convenient tabular form.
}
\examples{
\dontrun{
chart <- pollstr_chart_tbl()
chart
}
}

