% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts-csv.R
\name{pollstr_charts_tbl}
\alias{pollstr_charts_tbl}
\title{Pollster Charts (table)}
\usage{
pollstr_charts_tbl(slug)
}
\arguments{
\item{slug}{The slug of the chart, e.g. "012-general-election-romney-vs-obama.csv".}
}
\value{
A data frame with the chart data.
}
\description{
This provides a convenient table form
for the poll data from a Pollster Chart.
}

