% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pollstR-package.R
\docType{package}
\name{pollstR}
\alias{pollstR}
\alias{pollstR-package}
\title{R client for the Huffpost Pollster API}
\description{
This pacakge provides an R interface to the Huffington Post Pollster API.
Pollster provides programmatic access to opinion polls collected by the Huffington Post.
}
\details{
See \url{http://elections.huffingtonpost.com/pollster/api} for more details on the API.
}

